backend/requirements.txt
anyio==4.12.1
asgiref==3.11.1
beautifulsoup4==4.14.3
bota==4.0.155
botasaurus==4.0.97
botasaurus-humancursor==4.0.83
botasaurus-proxy-authentication==1.0.16
botasaurus-requests==4.0.38
botasaurus_api==4.0.10
botasaurus_driver==4.0.92
brotli==1.2.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.1.8
close-chrome==4.0.40
Deprecated==1.3.1
Django==4.2.29
django-cors-headers==4.9.0
django-filter==25.1
djangorestframework==3.16.1
et_xmlfile==2.0.0
exceptiongroup==1.3.1
gevent==25.9.1
geventhttpclient==2.3.9
greenlet==3.2.5
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
javascript-fixes==1.1.29
joblib==1.5.3
lxml==6.0.2
markdown-it-py==3.0.0
mdurl==0.1.2
numpy==2.0.2
openpyxl==3.1.5
packaging==26.0
playwright==1.58.0
psutil==7.2.2
psycopg2-binary==2.9.11
pyee==13.0.1
Pygments==2.19.2
pytweening==1.2.0
PyVirtualDisplay==3.0
requests==2.32.5
rich==14.3.3
soupsieve==2.8.3
sqlparse==0.5.5
typing_extensions==4.15.0
urllib3==2.6.3
websocket-client==1.9.0
wrapt==2.1.2
zope.event==6.0
zope.interface==8.0.1
